Refactor sorting script into general "reformat" script.
