+++ scripts/reformat-mappings.js
@@ -1,0 +1,46 @@
+const fs = require('fs');
+const mappings = require('../mappings');
+const reserved = require('../reserved');
+const { compare } = require('./shared');
+
+function sortByModuleAndExport(mappingA, mappingB) {
+  if (mappingA.module === mappingB.module) {
+    // ensure default exports sort higher within a package
+    let aExport = mappingA.export === 'default' ? '' : mappingA.export;
+    let bExport = mappingB.export === 'default' ? '' : mappingB.export;
+    return compare(aExport || '', bExport || '');
+  }
+
+  return compare(mappingA.module, mappingB.module);
+}
+
+let updated = mappings.sort(sortByModuleAndExport).map(mapping => {
+  if (!mapping.localName) {
+    mapping.localName = mapping.export;
+  }
+
+  if (reserved.includes(mapping.localName)) {
+    mapping.localName = 'Ember' + mapping.localName;
+  }
+
+  if (!('deprecated' in mapping)) {
+    mapping.deprecated = false;
+  }
+
+  // return a new object so that all of the properties in the JSON
+  // are in the same order
+  let ret = {
+    global: mapping.global,
+    module: mapping.module,
+    export: mapping.export,
+    localName: mapping.localName,
+    deprecated: mapping.deprecated,
+    replacement: mapping.replacement,
+  };
+
+  return ret;
+});
+
+fs.writeFileSync('mappings.json', JSON.stringify(updated, null, 2), {
+  encoding: 'utf-8',
+});
